move scripts again
