+++ src/scripts/v2/index.ts
