+++ src/scripts/v2/openapi/api.ts
@@ -1,11 +1,11 @@
-import details from '../../../src/schemas/v2/api/details'
-import routes from '../../../src/schemas/v2/api/paths'
-import * as components from '../../../src/schemas/v2/api/components'
+import details from '../../../schemas/v2/api/details'
+import routes from '../../../schemas/v2/api/paths'
+import * as components from '../../../schemas/v2/api/components'
 
-import { createQueryParameters } from '../../../src/schemas/v2/api/components/parameters'
-import { APIErrorCodes } from '../../../src/schemas/v2/api/components/errors'
+import { createQueryParameters } from '../../../schemas/v2/api/components/parameters'
+import { APIErrorCodes } from '../../../schemas/v2/api/components/errors'
 
-import { PatreonOauthScope } from '../../../src/rest/v2'
+import { PatreonOauthScope } from '../../../rest/v2'
 
 import { writeOpenAPISchema } from './schema'
 import { getJsDocDescription, getTypes } from '../shared'
@@ -51,8 +51,9 @@ export default () => writeOpenAPISchema({
             return {
                 documentation: {
                     description: 'Official documentation',
-                    url: `https://docs.patreon.com/#${method.toLowerCase()}-api-oauth2-v2${route.route(route.params?.id ?? 'id')
-                        .replace(/\//g, '-')}`,
+                    url: `https://docs.patreon.com/#${method.toLowerCase()}-api-oauth2-v2${
+                        route.route(route.params?.id ?? 'id').replace(/\//g, '-')
+                    }`,
                 },
                 responses: [
                     {
