+++ tsconfig.json
@@ -43,7 +43,6 @@
   ],
   "include": [
     "src/**/*.ts",
-    "scripts/**/*.ts",
     "tsup.config.ts"
   ]
 }
